pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roomestim"
version = "0.1.1"
description = "Capture-to-config tool: room scan -> RoomModel + speaker placement -> spatial_engine YAMLs"
requires-python = ">=3.10"
authors = [{ name = "paiiek" }]
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6",
    "jsonschema>=4",
    "shapely>=2.0",
    "scipy>=1.10",
    "trimesh>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-mock",
    "hypothesis>=6",
    "ruff>=0.5",
    "mypy>=1.8",
]
viz = ["matplotlib>=3.7"]
usd = ["pyusd"]
colmap = ["pycolmap>=0.6"]

[project.scripts]
roomestim = "roomestim.cli:main"

[tool.setuptools.packages.find]
include = ["roomestim*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
roomestim = ["proto/*.json"]

[tool.pytest.ini_options]
markers = [
    "lab: requires lab_real.usdz fixture; skipped in default CI",
]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
files = ["roomestim"]
